Move OttEvent to package com.kaltura.playkit.plugins.ott.
